Rename some key files and add a date converter
