+++ app/src/main/java/com/cvitirinyu/codingchallenge/starwars/adapters/BindingAdapters.kt
@@ -3,6 +3,7 @@ package com.cvitirinyu.codingchallenge.starwars.adapters
 import android.widget.ImageView
 import androidx.databinding.BindingAdapter
 import com.bumptech.glide.Glide
+import com.bumptech.glide.load.engine.DiskCacheStrategy
 import com.cvitirinyu.codingchallenge.starwars.R
 
 @BindingAdapter("imageFromUrl")
@@ -11,6 +12,7 @@ fun bindImageFromUrl(view: ImageView, imageUrl: String?) {
             .load(imageUrl)
             .centerCrop()
             .placeholder(R.drawable.placeholder_nomoon)
+            .diskCacheStrategy(DiskCacheStrategy.ALL)
             .into(view)
 
 }
