+++ app/src/main/java/com/cvitirinyu/codingchallenge/starwars/data/network/api/ApiInterface.kt
@@ -1,6 +1,6 @@
-package com.cvitirinyu.codingchallenge.starwars.data.api
+package com.cvitirinyu.codingchallenge.starwars.data.network.api
 
-import com.cvitirinyu.codingchallenge.starwars.data.model.StarWarsMissionResponse
+import com.cvitirinyu.codingchallenge.starwars.data.network.model.StarWarsMissionResponse
 import retrofit2.Response
 import retrofit2.http.GET
