+++ app/src/main/java/com/cvitirinyu/codingchallenge/starwars/data/network/model/StarWarsMissionResponse.kt
@@ -1,7 +1,5 @@
-package com.cvitirinyu.codingchallenge.starwars.data.model
+package com.cvitirinyu.codingchallenge.starwars.data.network.model
 
-import androidx.room.Entity
-import androidx.room.PrimaryKey
 import com.google.gson.annotations.SerializedName
 
 data class StarWarsMissionResponse (
